+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [UNRELEASED]
+### Added
+- Handle secured access (aka auth)
+
+### Fixed
+- Fix documentation
+
+### Updated
+- Use cleep config components
+
 ## [2.2.0] - 2023-03-10
 ### Changed
 - Clean code